requirements.txt
fastapi
pydantic[all]
email-validator
uvicorn
asyncpg
psycopg[binary]
passlib[bcrypt]
pyjwt
python-multipart
pytest
pytest-asyncio
databases
httpx
google-cloud-storage
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-aiplatform
celery[redis]
joblib==1.3.1
pycaret[full]
asyncio